requirements.txt
aiogram==3.24.0
aiosqlite
apscheduler
python-dotenv
gspread
google-auth

